Created and Implemented WaypointForm
- created a component for the WaypointForm
- decoupled the Waypoint Input form from the AutoNav page
- Fixed eslint errors
