+++ src/Components/Map.js
@@ -1,10 +1,8 @@
-import React, { useEffect, useState, Component, useCallback } from "react";
-import axios from "axios";
+import React, { useState, useCallback } from "react";
 import { MapContainer, TileLayer, Marker, Polyline } from 'react-leaflet'
 import 'leaflet/dist/leaflet.css';
 import L from 'leaflet';
-import { render } from "@testing-library/react";
-import { Button, Form } from "react-bootstrap";
+import { Button } from "react-bootstrap";
 
 
 // needed to properly display the default marker on the map
@@ -50,12 +48,6 @@ function Map(props) {
     const [roverLocation, setRoverLocation] = useState(state.location); //the current location of the rover
     const [objLocation, setObjLocation] = useState([]);
     const [lineLocation, setLineLocation] = useState([state.location]); //traces the motion of the rover
-    const [ waypoint, setWaypoint ] = useState({
-        type: "AR Tag",
-        position: [],
-        visited: false,
-        visible: false,
-    });
 
     // update locations of rover and trace line
     const updateLocation = useCallback((newLocation) => {
@@ -74,40 +66,7 @@ function Map(props) {
         setObjLocation([...objLocation, newLocation]);
     }
 
-    const handleChange = (e) => {
-        const value = e.target.value;
-        setWaypoint({
-            ...waypoint,
-            [e.target.name]: value,
-        });
-        console.log(waypoint);
-    };
 
-    const handleSubmit = (e) => {
-        const form = e.currentTarget;
-        e.preventDefault();
-        // if (form.checkValidity() === false) {
-        //     e.stopPropagation();
-        // }
-        // setValidated(true);
-        const newWaypoint = {
-            type: waypoint.type,
-            position: [Number(waypoint.latitude), Number(waypoint.longitude)],
-            visited: waypoint.visited,
-            visible: waypoint.visible,
-        };
-        if (!newWaypoint) {
-            console.log("xd");
-        } else {
-            axios
-                .post("http://localhost:9000/waypoints", newWaypoint)
-                .then((response) => {
-                    console.log(response.status);
-                    console.log(response.data.token);
-                    window.location.reload();
-                });
-        }
-    };
     function RoverMarker(props) {
 
         // updateLocation([21.2998, -157.8159]);
@@ -142,30 +101,6 @@ function Map(props) {
                 <Polyline positions={lineLocation} color="red" />
             </MapContainer>
             <Button onClick={() => {updateLocation(state.traceLine[3]); console.log(props.waypointData)}}>Next Destination</Button>
-            <h3 style={{ textAlign: "center" }}>Insert Waypoint</h3>
-            <Form onSubmit={handleSubmit}>
-                <Form.Select aria-label="Default select example">
-                    <option value="1">AR Tag</option>
-                    <option value="2">Goal</option>
-                </Form.Select>
-                <Form.Control
-                    type="number"
-                    step="0.0001"
-                    name="latitude"
-                    placeholder="enter latitude"
-                    onChange={handleChange}
-                    required
-                />
-                <Form.Control
-                    type="number"
-                    step="0.0001"
-                    name="longitude"
-                    placeholder="enter longitude"
-                    onChange={handleChange}
-                    required
-                />
-                <Button type="submit">Submit</Button>
-            </Form>
         </div>
     );
 }
